+++ packages/terra-framework-docs/src/terra-dev-site/doc/table/Examples/AutoLayoutDefaultTable.jsx
@@ -1,0 +1,43 @@
+import React from 'react';
+import Table from 'terra-table';
+
+const tableData = {
+  cols: [
+    { id: 'Column-0', displayName: 'Patient' },
+    { id: 'Column-1', displayName: 'Location' },
+    { id: 'Column-2', displayName: 'Illness Severity' },
+    { id: 'Column-3', displayName: 'Visit' },
+  ],
+  rows: [
+    {
+      id: '1',
+      cells: [
+        { content: 'Fleck, Arthur' },
+        { content: '1007-MTN' },
+        { content: 'Unstable' },
+        { content: 'Inpatient, 2 months' },
+      ],
+    },
+    {
+      id: '2',
+      cells: [
+        { content: 'Wayne, Bruce' },
+        { content: '1007-MTN-DR' },
+        { content: 'Stable' },
+        { content: 'Outpatient, 2 days' },
+      ],
+    },
+  ],
+};
+
+const AutoLayoutDefaultTable = () => (
+  <Table
+    id="my-table-id"
+    isAutoLayout
+    overflowColumns={tableData.cols}
+    rows={tableData.rows}
+    ariaLabel="Table"
+  />
+);
+
+export default AutoLayoutDefaultTable;